refactor(client-presence): rename internalRuntimeTypes.ts to runtimeTypes.ts
